analytic_api/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-redis==5.4.0
drf-spectacular==0.27.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
uvicorn==0.30.6
redis==5.0.8
pytest==8.3.3
pytest-django==4.8.0

